+++ CHANGELOG.md
@@ -1,12 +1,11 @@
 
 ### Added
 - Add a tutorial that describes how to deploy a Vitessce web app to GitHub pages.
+- Support `var` alias for AnnData to display altenrative gene names via new `geneAlias` field.
 
 ### Changed
 
 ## [1.1.19](https://www.npmjs.com/package/vitessce/v/1.1.19) - 2022-03-30
-
-
 
 ### Added
 - Added a roadmap page to the documentation.